WIP refactor code for reuse
